+++ src/gulp/spec.js
@@ -61,6 +61,7 @@ import yargs from 'yargs';
 import karma from 'karma';
 import istanbul from 'browserify-istanbul';
 import babelify from 'babelify';
+import tsify from 'tsify';
 import fs from 'fs';
 
 var argv = yargs.argv;
@@ -100,6 +101,8 @@ export default function(opts) {
     var stopAboveEslintThreshold = !!eslintThreshold;
     // coverage thresholds for each file
     var coverageThresholdEachFile = opts.coverageEachFile || {};
+    // if using typescript
+    var typescript = opts.typescript || false,
     // where the gulp task was ran from
     var originPath = process.cwd();
 
@@ -188,6 +191,10 @@ export default function(opts) {
     // Report tests slower than value
     if (argv['report-slow']) {
       config.reportSlowerThan = 100;
+    }
+
+    if (typescript) {
+      config.browserify.plugin = [ tsify ];
     }
 
     // tie the preprocessors to the relevant sources